refactor(codegen): extract tokens into plugin folder
