+++ scripts/codegen/src/core/registry.ts
@@ -1,4 +1,5 @@
 import { motionPlugin } from "../plugins/motion/index.ts";
+import { tokensPlugin } from "../plugins/tokens/index.ts";
 import type { SubstratePlugin } from "./plugin.ts";
 
 /**
@@ -6,4 +7,4 @@ import type { SubstratePlugin } from "./plugin.ts";
  * contributions are visible to later plugins' analyze() AND to every
  * plugin's emit(). Plugins never import one another.
  */
-export const PLUGINS: readonly SubstratePlugin[] = [motionPlugin];
+export const PLUGINS: readonly SubstratePlugin[] = [motionPlugin, tokensPlugin];
